move conversion_host configuration ops to be a convern
